% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_token.R
\name{ask_blink_api_token}
\alias{ask_blink_api_token}
\title{ask_blink_api_token}
\usage{
ask_blink_api_token(msg = "blink api token")
}
\arguments{
\item{msg}{the message}
}
\description{
ask for the blinke api token
}
