% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlinkR-package.R
\docType{package}
\name{BlinkR-package}
\alias{BlinkR-package}
\alias{BlinkR}
\title{renewr: Recovery of modulated renewal processes via Gaussian process modelling}
\description{
A compact toolkit for Gaussian-process modelling of time-series discrete event data
\if{html}{\figure{hex_sticker_RenewR.png}{options: style="float: right" width="150" alt="Hex Sticker"}}
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.7. https://mc-stan.org
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Caius-Gibeily/BlinkR}
  \item Report bugs at \url{https://github.com/Caius-Gibeily/BlinkR/issues}
}

}
\author{
Caius Gibeily \email{cgibeil@emory.edu}
}
