% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info}
\alias{get_info}
\title{getinfo}
\usage{
get_info(
  email,
  password,
  user_agent = "iPhone 9.2 | 2.2 | 222",
  unique_id = "00000000-0000-0000-0000-000000000000"
)
}
\arguments{
\item{email}{email}

\item{password}{password}
}
\description{
getinfo
}
