% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_networks.R
\name{get_networks}
\alias{get_networks}
\title{get_networks}
\usage{
get_networks(
  token = get_blink_api_token(),
  url = "https://rest.prde.immedia-semi.com/networks",
  host = "prde.immedia-semi.com"
)
}
\arguments{
\item{host}{}
}
\description{
get_networks
}
