% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_video.R
\name{get_video}
\alias{get_video}
\title{Title}
\usage{
get_video(videoPath, videoURL, token = get_blink_api_token(),
  host = "prde.immedia-semi.com")
}
\arguments{
\item{host}{}
}
\description{
Title
}
