% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_videos_link.R
\name{get_videos_link}
\alias{get_videos_link}
\title{get_video}
\usage{
get_videos_link(
  accountid,
  region,
  token = get_blink_api_token(),
  output_dir = "export",
  host = "prde.immedia-semi.com",
  max_pages = 1000
)
}
\description{
get_video
}
