% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_plot.R
\name{mcmc_plot}
\alias{mcmc_plot}
\title{Plot estimates from MCMC draws}
\usage{
mcmc_plot(
  model,
  pars = NULL,
  prior = FALSE,
  type = c("areas", "dens", "dens_chains", "hist", "intervals", "trace"),
  scheme = "blue",
  show_ground = TRUE,
  height = 0.3,
  linewidth = 1,
  color = "black",
  ...
)
}
\arguments{
\item{model}{gp_model. A fitted GP model or model with prior checks}

\item{pars}{Character or character vector. Set of parameters to plot, comprising
intercept parameters mu; GP parameters alpha and rho; and survival parameters k,
shape and sigma, if applicable. Characters may specify full parameter by name,
i.e. \code{"mu_group"} to plot all group intercepts or generically as \code{"mu"} to plot
all intercepts. By default, all parameters are shown.}

\item{prior}{Boolean. Plot prior estimates, if available, instead of posterior estimates.
The default is \code{FALSE}}

\item{type}{Character. Choice of output visualisation from \code{bayesplot::mcmc_} function set.
The type can be \code{"areas"} (see \code{\link[bayesplot:mcmc_areas]{bayesplot::mcmc_areas()}}), \code{"dens"} (see \code{\link[bayesplot:mcmc_dens]{bayesplot::mcmc_dens()}}),
\code{"dens_chains"} (see \code{\link[bayesplot:mcmc_dens_chains]{bayesplot::mcmc_dens_chains()}}), \code{"hist"} (see \code{\link[bayesplot:mcmc_hist]{bayesplot::mcmc_hist()}}),
\code{"intervals"} (see \code{\link[bayesplot:mcmc_intervals]{bayesplot::mcmc_intervals()}}) and \code{"trace"} (see \code{\link[bayesplot:mcmc_trace]{bayesplot::mcmc_trace()}}).
The default is \code{"intervals"}.}

\item{scheme}{Character. Colour scheme of plot. Option is passed to \code{\link[bayesplot:color_scheme_set]{bayesplot::color_scheme_set()}}.
The default is \code{"blue"}.}

\item{show_ground}{Boolean. If the model was fitted on simulated ground truth data, i.e.
\code{\link[=simulate_gp_traces]{simulate_gp_traces()}}, setting \code{show_ground=TRUE} will show the respective parameters
underlying the simulated data as vertical lines. By default, this is true but will
not raise an error ground truth data are not available.}

\item{height}{Numeric. Height of the vertical line signposting the ground truth
parameter values if \code{show_ground=TRUE}.}

\item{linewidth}{Numeric. Width of the vertical line signposting the ground truth parameter
values if \code{show_ground=TRUE}.}

\item{color}{Character. The colour of the vertical line signposting the ground truth parameter
values if \code{show_ground=TRUE}.}

\item{...}{Optional arguments passed to the selected mcmc_* plotting function.}
}
\description{
A wrapper function on \strong{bayesplot} to simplify visualisation of prior or posterior estimates from
MCMC draws.
}
\examples{
# Simulate traces
gp_traces <- simulate_gp_traces(n_ind = 1, alpha_ind = 0.5,
    rho_ind = 10,
    seed = 123)
# Simulate events
events <- simulate_events(gp_traces, family = "weibull", baseline = 0.5)

# Fit model

fit <- gp_fit(events, family = "gamma",
    priors = alist(alpha_ind ~ normal(0.5, 0.3),
    rho_ind ~ lognormal(2.2, 0.4),
    mu_ind ~ normal(0.5, 0.2),
    shape ~ normal(1.5, 0.3)
    ))
# Plot mcmc intervals of the core model parameters
mcmc_plot(fit)
# Plot densities by chain to check chain convergence, setting red colour scheme
mcmc_plot(fit, type = "dens_chains", scheme = "red")
# Remove ground truth lines from plot
mcmc_plot(fit, show_ground = FALSE)
}
\seealso{
\code{\link[bayesplot:mcmc_areas]{bayesplot::mcmc_areas()}}, \code{\link[bayesplot:mcmc_dens]{bayesplot::mcmc_dens()}},
\code{\link[bayesplot:mcmc_dens_chains]{bayesplot::mcmc_dens_chains()}}, \code{\link[bayesplot:mcmc_hist]{bayesplot::mcmc_hist()}},
\code{\link[bayesplot:mcmc_intervals]{bayesplot::mcmc_intervals()}}) and \code{\link[bayesplot:mcmc_trace]{bayesplot::mcmc_trace()}}
}
