% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_plot.R
\name{plot.sim_traces}
\alias{plot.sim_traces}
\title{Plot Simulation Traces}
\usage{
\method{plot}{sim_traces}(
  sim_data,
  level = c("individual", "group", "global"),
  facet = TRUE,
  separate_pages = FALSE,
  width = 0.1,
  height = 0.2,
  size = 10,
  ...
)
}
\arguments{
\item{level}{Character vector specifying which levels to plot.
Options include \code{"individual"}, \code{"group"}, or \code{"global"}. Defaults to all three.}

\item{facet}{Logical. If \code{TRUE}, splits plots up by group or individual.}

\item{separate_pages}{Logical. If \code{TRUE} and \code{facet = TRUE}, outputs a layout
across multiple pages via \code{gridExtra}.}

\item{...}{Additional arguments passed to methods (ignored).}

\item{x}{An object of class \code{sim_traces}.}
}
\description{
Plot Simulation Traces
}
