% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_checks.R
\name{ppc_get_eventrate}
\alias{ppc_get_eventrate}
\title{Plot overall event rate from prior or posterior predictive distribution}
\usage{
ppc_get_eventrate(
  model,
  ppc_events = NULL,
  scale_factor = 1,
  .width = c(0.5, 0.8, 0.99),
  palette = "Purples",
  return_plot = TRUE,
  n_samples = 1000,
  prior = FALSE
)
}
\arguments{
\item{model}{gp_model. A fitted GP model or model with prior checks}

\item{ppc_events}{tibble. Optional tibble of pre-drawn prior/posterior predictive
events generated via \code{\link[=ppc_draw_events]{ppc_draw_events()}}. If ppc_events is \code{NULL}, the default,
\code{model} cannot be \code{NULL} and events will be generated via the function on the passed model.}

\item{scale_factor}{Numeric. The factor to scale event rate and instantaneous
event rate, both of which are by default in events/s.}

\item{.width}{Numeric vector. Highest density posterior credible interval bands around
the mean GP trace to visualise. Strictly, widths show pointwise quantiles around
the mean. By default, \code{.width = c(0.5, 0.8, 0.99)}, marking 50\%, 80\% and 99\% density
coverage (see \code{\link[ggdist:interval_widths]{ggdist::interval_widths()}}).}

\item{palette}{Character. Colour scheme for plotting credible intervals. The choice
of palette is passed to \code{\link[ggplot2:scale_fill_brewer]{ggplot2::scale_fill_brewer()}} (see \code{\link[RColorBrewer:display.brewer.all]{RColorBrewer::display.brewer.all()}} for
the full set of colour palettes).}

\item{return_plot}{Boolean. Specify whether to return a ggplot or a list of data.
By default, this is TRUE}

\item{n_samples}{Integer. Number of samples to draw to compute prior/posterior
predictive checks. By default this is 1000 and can be no greater than the number of
post-warmup iterations times the number of chains (see \code{\link[=gp_fit]{gp_fit()}}). Using the default
parameters means this ceiling is 4000 samples.}

\item{prior}{Boolean. Plot prior estimates, if available, instead of posterior estimates.
The default is \code{FALSE}}
}
\value{
Either a ggplot or a list of observed
overall blink rates per individual and blink rate distributions via PPC.
}
\description{
Plot overall event rate from prior or posterior predictive distribution
}
\seealso{
\code{\link[=ppc_get_interevent_dist]{ppc_get_interevent_dist()}}, \code{\link[=ppc_get_inst_eventrate]{ppc_get_inst_eventrate()}}, \code{\link[=ppc_plot_all]{ppc_plot_all()}}.
}
