% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_checks.R
\name{ppc_get_inst_eventrate}
\alias{ppc_get_inst_eventrate}
\title{Plot instantaneous event rates from prior or posterior predictive distributions}
\usage{
ppc_get_inst_eventrate(
  model,
  ppc_events = NULL,
  bw_rate = 8,
  resolution = 0.01,
  n_samples = 1000,
  .width = c(0.5, 0.8, 0.99),
  return_plot = TRUE,
  palette = "Purples",
  show_events = TRUE,
  prior = FALSE
)
}
\arguments{
\item{model}{gp_model. A fitted GP model or model with prior checks}

\item{ppc_events}{tibble. Optional tibble of pre-drawn prior/posterior predictive
events generated via \code{\link[=ppc_draw_events]{ppc_draw_events()}}. If ppc_events is \code{NULL}, the default,
\code{model} cannot be \code{NULL} and events will be generated via the function on the passed model.}

\item{bw_rate}{Numeric. The smoothing bandwidth to use for
computing the instantaneous event rate}

\item{resolution}{Numeric. The resolution of the time grid for computing instantaneous
event rate (see \code{\link[=ppc_get_inst_eventrate]{ppc_get_inst_eventrate()}}). By default, this is 0.01}

\item{n_samples}{Integer. Number of samples to draw to compute prior/posterior
predictive checks. By default this is 1000 and can be no greater than the number of
post-warmup iterations times the number of chains (see \code{\link[=gp_fit]{gp_fit()}}). Using the default
parameters means this ceiling is 4000 samples.}

\item{.width}{Numeric vector. Highest density posterior credible interval bands around
the mean GP trace to visualise. Strictly, widths show pointwise quantiles around
the mean. By default, \code{.width = c(0.5, 0.8, 0.99)}, marking 50\%, 80\% and 99\% density
coverage (see \code{\link[ggdist:interval_widths]{ggdist::interval_widths()}}).}

\item{return_plot}{Boolean. Specify whether to return a ggplot or a list of data.
By default, this is TRUE}

\item{palette}{Character. Colour scheme for plotting credible intervals. The choice
of palette is passed to \code{\link[ggplot2:scale_fill_brewer]{ggplot2::scale_fill_brewer()}} (see \code{\link[RColorBrewer:display.brewer.all]{RColorBrewer::display.brewer.all()}} for
the full set of colour palettes).}

\item{show_events}{Boolean. Determines whether event data, if available are shown
as subplots beneath each plot of GP traces. By default, if event data are available,
they are plotted (\code{show_events=TRUE}).}

\item{prior}{Boolean. Plot prior estimates, if available, instead of posterior estimates.
The default is \code{FALSE}}
}
\value{
A ggplot or a list of data
}
\description{
Plot instantaneous event rates from prior or posterior predictive distributions
}
\seealso{
\code{\link[=ppc_get_interevent_dist]{ppc_get_interevent_dist()}}, \code{\link[=ppc_get_eventrate]{ppc_get_eventrate()}}, \code{\link[=ppc_plot_all]{ppc_plot_all()}}.
}
