% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_checks.R
\name{ppc_plot_all}
\alias{ppc_plot_all}
\title{Plot prior or posterior predictive checks on fitted model}
\usage{
ppc_plot_all(
  model,
  n_samples = 1000,
  palette = "Blues",
  prior = FALSE,
  .width = c(0.5, 0.8, 0.99),
  bw_kde = 0.5,
  bw_rate = 5,
  scale_factor = 1,
  resolution = 0.01,
  ...
)
}
\arguments{
\item{model}{gp_model. A fitted GP model or model with prior checks}

\item{n_samples}{Integer. Number of samples to draw to compute prior/posterior
predictive checks. By default this is 1000 and can be no greater than the number of
post-warmup iterations times the number of chains (see \code{\link[=gp_fit]{gp_fit()}}). Using the default
parameters means this ceiling is 4000 samples.}

\item{palette}{Character. Colour scheme for plotting credible intervals. The choice
of palette is passed to \code{\link[ggplot2:scale_fill_brewer]{ggplot2::scale_fill_brewer()}} (see \code{\link[RColorBrewer:display.brewer.all]{RColorBrewer::display.brewer.all()}} for
the full set of colour palettes).}

\item{prior}{Boolean. Plot prior estimates, if available, instead of posterior estimates.
The default is \code{FALSE}}

\item{.width}{Numeric vector. Highest density posterior credible interval bands around
the mean GP trace to visualise. Strictly, widths show pointwise quantiles around
the mean. By default, \code{.width = c(0.5, 0.8, 0.99)}, marking 50\%, 80\% and 99\% density
coverage (see \code{\link[ggdist:interval_widths]{ggdist::interval_widths()}}).}

\item{bw_kde}{Numeric or character. The smoothing bandwidth to use for the interevent interval
distribution. The default is \code{"nrd0"} (see \code{\link[stats:density]{stats::density()}})}

\item{bw_rate}{Numeric. The smoothing bandwidth to use for
computing the instantaneous event rate}

\item{scale_factor}{Numeric. The factor to scale event rate and instantaneous
event rate, both of which are by default in events/s.}

\item{resolution}{Numeric. The resolution of the time grid for computing instantaneous
event rate (see \code{\link[=ppc_get_inst_eventrate]{ppc_get_inst_eventrate()}}). By default, this is 0.01}

\item{...}{Optional plotting arguments.}
}
\value{
A series of ggplots.
}
\description{
Plot a series of posterior predictive checks based on \code{\link[=ppc_get_eventrate]{ppc_get_eventrate()}},
\code{\link[=ppc_get_interevent_dist]{ppc_get_interevent_dist()}} and \code{\link[=ppc_get_inst_eventrate]{ppc_get_inst_eventrate()}}.
}
