% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gp_model.R
\name{print.gp_model}
\alias{print.gp_model}
\title{Print table of inferred parameters}
\usage{
\method{print}{gp_model}(
  model,
  prior = FALSE,
  pars = NULL,
  probs = c(0.25, 0.5, 0.75),
  digits_summary = 2,
  ...
)
}
\arguments{
\item{model}{gp_model. A fitted GP model or model with prior checks}

\item{prior}{Boolean. Determines whether to show GPs drawn from prior or posterior
parameters. By default, \code{prior=FALSE} (see \code{\link[=gp_fit]{gp_fit()}}).}

\item{probs}{Numeric or numeric vector. Parameter quantiles to print. By default
\code{probs=c(0.25,0.5,0.75)}, printing 25th, 50th and 75th percentile parameter values}

\item{digits_summary}{Integer. Number of significant digits to round to. By default,
this is 2.}

\item{...}{Other optional arguments}
}
\description{
Print table of inferred parameters
}
