% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_traces.R
\name{reconstruct_traces}
\alias{reconstruct_traces}
\title{Reconstruct traces from Hilbert space}
\usage{
reconstruct_traces(
  model,
  level = c("ind", "group", "global"),
  prior = FALSE,
  dev_only = FALSE,
  resolution = 0.1
)
}
\arguments{
\item{model}{gp_model. A fitted GP model or model with prior checks}

\item{level}{Character. The level (\code{"ind"}, \code{"group"}, \code{"global"}) at which
the posterior fitted Gaussian Processes are extracted. By default, the level is \code{"ind"},
thereby showing the additive combination of group-level and individual GPs translated by
their respective baselines, \code{mu_ind[i]}, where i is the individual ID.}

\item{prior}{Boolean. Determines whether to show GPs drawn from prior or posterior
parameters. By default, \code{prior=FALSE} (see \code{\link[=gp_fit]{gp_fit()}}).}

\item{dev_only}{Boolean. Determines whether only the deviations relative to
the level above are to be plotted. For example, if \code{dev_only=TRUE} and \code{level="ind"}
in a single-group model, only \eqn{\mu_{ind[i]} + f_{ind[i]}} would be shown
for the ith individual, rather than the composite \eqn{\mu_{ind[i]} + f_{ind[i]} + f_{group[i]}}.}
}
\description{
Reconstruct traces from Hilbert space
}
\seealso{
\code{\link[=tidy_traces]{tidy_traces()}}, \code{\link[=draw_traces]{draw_traces()}}, \code{\link[=summarise_traces]{summarise_traces()}},
\code{\link[=reconstruct_traces]{reconstruct_traces()}}
}
