% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_token.R
\name{set_blink_api_token}
\alias{set_blink_api_token}
\title{set_blink_api_token}
\usage{
set_blink_api_token(token)
}
\arguments{
\item{token}{blink api token}
}
\description{
set the blink api token
}
