% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trajectories.R
\name{simulate_boxcar_traces}
\alias{simulate_boxcar_traces}
\title{Simulate Boxcar Trajectories}
\usage{
simulate_boxcar_traces(
  duration = 100,
  n_ind = 10,
  n_groups = 1,
  nodes = seq(1, duration, length.out = 3),
  steps = c(-1, 1),
  cycles = 2,
  sd_group = 0.5,
  sd_ind = 0.25
)
}
\arguments{
\item{duration}{Integer. The length of the full time series to simulate (in seconds). The default is 100 seconds}

\item{n_ind}{Integer or integer vector. Number of individuals per group. If 1 integer is supplied,}

\item{n_groups}{Integer. The number of groups to simulate. If the number of groups is greater than 1,
group splines are constrained to sum to zero. The default is 1 group.}

\item{nodes}{Numeric vector. Node locations for step intervals.}

\item{steps}{Numeric vector. Step sizes for the boxcar transitions.}

\item{cycles}{Integer. Number of repeating cycles over the duration.}

\item{sd_group}{Numeric. Group-level step scaling standard deviation.}

\item{sd_ind}{Numeric. Individual-level step scaling standard deviation.}
}
\value{
An object of class \code{sim_traces}, containing a list of \code{ind_traces},
\code{group_traces} (trace common to individuals in the same group),
\code{global_traces} (trace common to all individuals; NULL if simulation has only
1 individual or 1 group) and \code{sim_parameters} (simulation parameters).
}
\description{
Simulate Boxcar Trajectories
}
\examples{
boxcar_traces <- simulate_boxcar_traces(n_ind = 5, n_groups = 2)
plot(boxcar_traces)
}
\seealso{
The methods \code{\link[=simulate_spline_traces]{simulate_spline_traces()}} and
\code{\link[=simulate_gp_traces]{simulate_gp_traces()}}
}
