% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trajectories.R
\name{simulate_spline_traces}
\alias{simulate_spline_traces}
\title{Simulate Latent Spline Traces}
\usage{
simulate_spline_traces(
  duration = 100,
  n_ind = 10,
  n_groups = 1,
  df = 5,
  sd_group = 0.5,
  sd_ind = 0.25,
  mu_coeffs = stats::rnorm(df, 0, 1),
  seed = NULL
)
}
\arguments{
\item{duration}{Integer. The length of the full time series to simulate (in seconds). The default is 100 seconds}

\item{n_ind}{Integer or integer vector. Number of individuals per group. If 1 integer is supplied,}

\item{n_groups}{Integer. The number of groups to simulate. If the number of groups is greater than 1,
group splines are constrained to sum to zero. The default is 1 group.}

\item{df}{Integer. Degrees of freedom for the cubic spline basis set. The default is 5}

\item{sd_group}{Numeric. Standard deviation for group-level coefficients.
If the number of individuals is greater than 1, the s.d. of the group-level coefficients determines how  standard deviation.}

\item{sd_ind}{Numeric. Individual-level coefficient standard deviation.}

\item{mu_coeffs}{Numeric vector. Mean coefficients for the spline basis.}

\item{seed}{Numeric. Set a seed for reproducibility. If one is not specified, seed defaults to NULL}
}
\value{
An object of class \code{sim_traces}, containing a list of \code{ind_traces},
\code{group_traces} (trace common to individuals in the same group),
\code{global_traces} (trace common to all individuals; NULL if simulation has only
1 individual or 1 group) and \code{sim_parameters} (simulation parameters).
}
\description{
A short description...
}
\examples{
spline_traces <- simulate_spline_traces(n_ind = 5, n_groups = 2)
plot(spline_traces)

}
\seealso{
The methods \code{\link[=simulate_gp_traces]{simulate_gp_traces()}} and \code{\link[=simulate_boxcar_traces]{simulate_boxcar_traces()}}
}
