% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_traces.R
\name{summarise_traces}
\alias{summarise_traces}
\title{Summarise mean and standard deviation envelopes of recovered GP traces}
\usage{
summarise_traces(
  model = NULL,
  recon_data = NULL,
  level = "ind",
  prior = FALSE,
  dev_only = FALSE,
  resolution = 0.1,
  rescale = FALSE
)
}
\arguments{
\item{model}{gp_model. A fitted GP model or model with prior checks}

\item{recon_data}{Optional. Reconstructed trace data created by \code{\link[=reconstruct_traces]{reconstruct_traces()}}.
If NULL, then \code{model} cannot be NULL; the traces will first be reconstructed and then
transformed into tidy format.}

\item{level}{Character. The level (\code{"ind"}, \code{"group"}, \code{"global"}) at which
the posterior fitted Gaussian Processes are extracted. By default, the level is \code{"ind"},
thereby showing the additive combination of group-level and individual GPs translated by
their respective baselines, \code{mu_ind[i]}, where i is the individual ID.}

\item{prior}{Boolean. Determines whether to show GPs drawn from prior or posterior
parameters. By default, \code{prior=FALSE} (see \code{\link[=gp_fit]{gp_fit()}}).}

\item{dev_only}{Boolean. Determines whether only the deviations relative to
the level above are to be plotted. For example, if \code{dev_only=TRUE} and \code{level="ind"}
in a single-group model, only \eqn{\mu_{ind[i]} + f_{ind[i]}} would be shown
for the ith individual, rather than the composite \eqn{\mu_{ind[i]} + f_{ind[i]} + f_{group[i]}}.}

\item{rescale}{Boolean. The observed traces represent the time-modulated scale, \eqn{\theta(t)},
of the renewal process. However, for non-memoryless processes, namely, the exponential
the expected inter-event interval is also specified by shape parameters. \code{rescale}
applies a scaling constant, \eqn{c}, to the fitted GP traces (and simulated ground truth traces if applicable)
that corresponds with the survival family.
\deqn{\mathbb{E}[X] = \exp{\beta(t)} c \text{, where } \beta(t) = \ln{\theta(t)}}
By default, the traces are shown on the response scale and hence the inverse ln
is applied to \eqn{c}:
\deqn{\mathbb{E}[X] = \beta(t) + \ln{c}}
Where for the gamma family, \eqn{c = k}.
For Weibull, \eqn{c = \Gamma(1 + \frac{1}{\text{shape}})}.
For the generalised gamma, \eqn{c = \frac{\frac{\Gamma(k + 1)}{\Gamma(shape)}}{\frac{\Gamma(k)}{\Gamma(shape)}}}
Rescaling ensures that the expected inter-event interval (accounting for scale and shape)
is shown. For \code{\link[=get_nlpd]{get_nlpd()}}, rescale is by default \code{TRUE}.}
}
\value{
Summary trace data, showing mean and standard deviation at each sampled
time point along each recovered trace.
}
\description{
Summarise mean and standard deviation envelopes of recovered GP traces
}
\seealso{
\code{\link[=tidy_traces]{tidy_traces()}}, \code{\link[=draw_traces]{draw_traces()}}, \code{\link[=reconstruct_traces]{reconstruct_traces()}}
}
