% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlinkR-package.R
\docType{package}
\name{BlinkR-package}
\alias{BlinkR-package}
\alias{BlinkR}
\title{The 'BlinkR' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.7. https://mc-stan.org
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Caius-Gibeily/BlinkR}
  \item Report bugs at \url{https://github.com/Caius-Gibeily/BlinkR/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

Authors:
\itemize{
  \item First Last \email{first.last@example.com}
}

}
