% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_fit.R
\name{gp_fit}
\alias{gp_fit}
\title{Simulate inhomogeneous renewal processes}
\usage{
gp_fit(traces, ...)
}
\arguments{
\item{sigma}{Numeric. Individual-level coefficient standard deviation.}

\item{Q}{Numeric. Mean coefficients for the spline basis.}

\item{baseline}{Numeric. Group-level coefficient standard deviation.}

\item{baseline_sd}{Numeric. Individual-level coefficient standard deviation.}

\item{lerp}{Numeric. Mean coefficients for the spline basis.}
}
\value{
A tibble containing simulated event times for each group.
}
\description{
Simulate inhomogeneous renewal processes
}
