% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_.R
\name{mcmc_plot}
\alias{mcmc_plot}
\title{Title}
\usage{
mcmc_plot(
  model,
  pars = NULL,
  prior = FALSE,
  type = c("areas", "dens", "dens_chains", "hist", "intervals", "trace"),
  scheme = "blue",
  show_ground = TRUE,
  height = 0.3,
  linewidth = 1,
  color = "black",
  ...
)
}
\description{
Title
}
