% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sim_traces.R
\name{plot.sim_traces}
\alias{plot.sim_traces}
\title{Plot Traces of Simulated Data}
\usage{
\method{plot}{sim_traces}(
  sim_data,
  level = c("individual", "group", "global"),
  facet = TRUE,
  width = 0.1,
  height = 0.2,
  size = 10
)
}
\arguments{
\item{sim_data}{An object of class \code{sim_traces} or \code{sim_events}.}

\item{level}{Character vector specifying which levels to plot.
Options include \code{"ind"} (individual), \code{"group"}, or \code{"global"}. Defaults to all three.}

\item{facet}{Logical. If \code{TRUE}, splits plots into separate panels by group or
individual.}

\item{width}{Numeric. Width of the event rasters. The default is 0.1. Passed to \code{\link[ggplot2:geom_tile]{ggplot2::geom_tile()}}}

\item{height}{Numeric. Height of the event rasters. The default is 0.2. Passed to \code{\link[ggplot2:geom_tile]{ggplot2::geom_tile()}}}

\item{size}{Numeric. Axis label size. Default is 10}
}
\value{
A ggplot showing plotted traces at the level(s) specified and event raster
subplots faceted by individual or group, if specified.
}
\description{
Plot Traces of Simulated Data
}
\examples{
gp_traces <- simulate_gp_traces(n_ind = 5, alpha_group = 0.5, rho_group = 10,
seed = 123)
# Plot traces
plot(gp_traces)

# Simulate a modulated point process drawn from a Weibull and plot the traces + events
events <- simulate_events(gp_traces, family = "weibull")
plot(events)

}
