% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_events.R
\name{simulate_events}
\alias{simulate_events}
\title{Simulate inhomogeneous renewal processes}
\usage{
simulate_events(
  trace_data,
  group = group,
  ind = ind,
  family = c("exponential", "gamma", "weibull", "log-normal", "gengamma"),
  shape = 1,
  k = 2,
  sigma = 1,
  Q = 0,
  baseline = 0,
  lerp = 100,
  seed = NULL
)
}
\description{
Simulate inhomogeneous renewal processes
}
