% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trajectories-PED-46RVV44.R,
%   R/simulate_trajectories.R
\name{simulate_gp_trajectories}
\alias{simulate_gp_trajectories}
\title{Simulate Gaussian Process Trajectories}
\usage{
simulate_gp_trajectories(
  duration = 100,
  n_groups = 1,
  n_ind = 10,
  kernel = "squared_exp",
  alpha_global = 0.5,
  rho_global = duration/2,
  alpha_group = 0.5,
  rho_group = duration/2,
  alpha_ind = 0.25,
  rho_ind = duration/2,
  nu = 3/2,
  P = 1
)

simulate_gp_trajectories(
  duration = 100,
  n_groups = 1,
  n_ind = 10,
  kernel = "squared_exp",
  alpha_global = 0.5,
  rho_global = duration/2,
  alpha_group = 0.5,
  rho_group = duration/2,
  alpha_ind = 0.25,
  rho_ind = duration/2,
  nu = 3/2,
  P = 1
)
}
\arguments{
\item{duration}{Integer. The length of the time series.}

\item{n_groups}{Integer. Number of groups.}

\item{n_ind}{Integer. Number of individuals.}

\item{kernel}{Character. Type of GP kernel ("squared_exp", "matern", "rational_quad", "periodic").}

\item{alpha_global, alpha_group, alpha_ind}{Numeric. Covariance amplitude parameters.}

\item{rho_global, rho_group, rho_ind}{Numeric. Length-scale parameters.}

\item{nu}{Numeric. Matern smoothness parameter.}

\item{P}{Numeric. Period duration parameter for periodic kernels.}
}
\value{
A list containing \code{group_traces} and \code{ind_traces}.

A list containing \code{group_traces} and \code{ind_traces}.
}
\description{
Simulate Gaussian Process Trajectories

Simulate Gaussian Process Trajectories
}
