% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trajectories.R
\name{simulate_spline_trajectories}
\alias{simulate_spline_trajectories}
\title{Simulate Latent Trajectories (splines, gaussian processes and boxcars)}
\usage{
simulate_spline_trajectories(
  duration = 100,
  n_ind = 10,
  n_groups = 1,
  df = 5,
  sd_group = 0.5,
  sd_ind = 0.25,
  mu_coeffs = stats::rnorm(df, 0, 1),
  seed = NULL
)
}
\arguments{
\item{duration}{Integer. The length of the time series.}

\item{n_ind}{Integer. Number of individuals.}

\item{n_groups}{Integer. Number of groups.}

\item{df}{Integer. Degrees of freedom for the cubic spline basis.}

\item{sd_group}{Numeric. Group-level coefficient standard deviation.}

\item{sd_ind}{Numeric. Individual-level coefficient standard deviation.}

\item{mu_coeffs}{Numeric vector. Mean coefficients for the spline basis.}
}
\value{
A list containing \code{group_traces} and \code{ind_traces}.
}
\description{
Simulate Latent Trajectories (splines, gaussian processes and boxcars)
}
